fin_app/requirements.txt
flask
requests
firebase-admin
google-cloud-pubsub
google-cloud-vision